requirements.txt
Flask
concrete-ml>=0.3.0
numpy
pandas
scikit-learn
requests
cryptography
pysqlcipher3
waitress